Make `AddOnAuto` client testable
Avoiding the execution of `#main` method after class definition, which
is a legacy behavior from the conversion from YCP language to Ruby.
